Document and add comments to the C‑API code. Create a .toml installer that supports multiple Linux platforms and various Python versions. Replace Vec<map> with VecMapNodes for faster lookup of the node’s localId within a partition.

+++ manapy/backends/types.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from mpi4py import MPI
 
-FLOAT_TYPE = "float64"
+FLOAT_TYPE = "float32"
 INT_TYPE = "int32"
 
 if FLOAT_TYPE not in ["float32", "float64"]:
